Swapped out recursive subdivide sphere for a latitude/longitude generated model.
Calculated normals for sphere to retain lighting from previous sphere model.

Changed bacteria to render as spheres rendering on the surface of the main sphere.

Added scalor and translation matrices to redraw the main sphere as bacteria.

Added velocity, acceleration and drag force to the rotation of the sphere, so it speeds up and slows down when force is applied instead of spinning at a constant rate.